# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexGuardian Foundation API",
    "version": "0.1.0",
    "description": "Contrato proposto, sem backend. Nenhum endpoint do produto original é utilizado."
  },
  "servers": [
    {
      "url": "https://api.nexguardian.invalid/v1",
      "description": "Placeholder não operacional; substituir por infraestrutura própria."
    }
  ],
...
  "paths": {
...
    "/account/me": {
      "get": {
        "operationId": "getAccount",
        "tags": [
          "Account"
        ],
        "summary": "Dados da conta autenticada",
        "description": "Escopo de conta pelo token; sem identificador em query.",
        "security": [
          {
            "accountBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountMeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Proibido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Muitas requisições",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/activation/codes": {
      "post": {
        "operationId": "issueActivationCode",
        "tags": [
          "Activation"
        ],
        "summary": "Emite um novo código de ativação",
        "description": "Escopo de conta pelo token.",
        "security": [
          {
            "accountBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Escopo por principal/operação."
          }
        ],
        "responses": {
          "201": {
            "description": "Criado",
            "headers": {
              "Cache-Control": {
                "description": "no-store",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueActivationCodeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pedido inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Proibido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Muitas requisições",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST",
              "INVALID_CREDENTIALS",
              "ACTIVATION_INVALID",
              "ACTIVATION_EXPIRED",
              "ACTIVATION_CONSUMED",
              "PAIRING_EXPIRED",
              "PAIRING_STATE_CONFLICT",
              "SESSION_REVOKED",
              "REFRESH_REUSED",
              "IDEMPOTENCY_CONFLICT",
              "FORBIDDEN",
              "NOT_FOUND",
              "RATE_LIMITED",
              "SERVER_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
...
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          }
        },
        "required": [
          "id",
          "accountId",
          "name",
          "email"
        ]
      },
...
      "AccountMeResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "account": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "profile": {
                "type": "string",
                "enum": [
                  "INDIVIDUAL",
                  "FAMILY",
                  "BUSINESS"
                ]
              }
            },
            "required": [
              "id",
              "name",
              "profile"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "deviceCount": {
            "type": "integer",
            "minimum": 0
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "account",
          "user",
          "deviceCount",
          "serverTime"
        ]
      },
...
      "IssueActivationCodeResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "value": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "value",
          "expiresAt",
          "serverTime"
        ]
      }
    },
    "securitySchemes": {
      "accountBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access token de conta. O escopo de isolamento (accountId, que mapeia o Workspace do domínio — ver ADR-0003) é transportado como claim no token; nenhum endpoint aceita identificador de conta/workspace por query ou path."
      },
      "deviceBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Access token de sessão de dispositivo, com audiência de dispositivo, vinculado a deviceId e ao escopo de conta/workspace (accountId) transportado como claim no token; nenhum endpoint aceita esse identificador por query ou path."
      },
      "activationTicket": {
        "type": "http",
        "scheme": "bearer",
        "description": "Ticket temporário exclusivo de validate → pair, ligado à conta e ao UUID da instalação."
      },
      "pairingTicket": {
        "type": "http",
        "scheme": "bearer",
        "description": "Ticket temporário exclusivo para confirmar um pareamento pendente."
      }
    }
  }
}